windows_app_lock/requirements.txt
# Core dependencies (required)
psutil>=5.8.0

# Optional dependencies for enhanced features
# System tray functionality
Pillow>=8.0.0
pystray>=0.19.0

# Windows-specific features
win10toast>=0.9; sys_platform == "win32"
pywin32>=300; sys_platform == "win32"

# Alternative packages for different platforms
# For Linux system tray (alternative to pystray)
# plyer>=2.0.0; sys_platform == "linux"